[FIX] sale_tier_validation: Basic compatibility with sale_loyalty
If you install sale_loyalty, this piece of code:

https://github.com/odoo/odoo/blob/2a7876538e9ea630563e39ee8402b27147d1e428/addons/sale_loyalty/models/sale_order.py#L740

is always executed when confirming a sales order, so it blocks it, no
matter if no coupons are being handled.

This little code avoids this blocking without requiring a glue module,
although further work would be needed if some coupons are applied.

TT51484