[MIG] sale_order_weight: Migration to 16.0 (from 12.0)
[FIX] the weight on sale.order.line is now always correct.

Change weight on product now change automatically the weight of the sale orders.
So it is not necessary to 'recompute' the weight of the orders,
and the value is always correct.
(Also the column 'Unit Weight' is removed at sale.order.line level)

[FIX] Handle correctly UoM conversion.

Use Case :
- product.product (weight : 50kg ; Uom : Unit)
- sale.order.line (Qty : 10 ; UoM : dozen)
-> Before, the weight was 50 * 10 = 500 and was wrong.
-> Now the weight is 50 * 10 * 12 = 6000

[MIG] Add script to fix incorrect weight values

The values can be incorrect (due to conversion errors),
or obsolet (if user didn't click on 'recompute weight')

[IMP] Display weight on sale order report.

Note: this code comes from the other OCA modules name sale_order_weight,
present in the OCA repo sale-reporting, in V14:
See: https://github.com/OCA/sale-reporting/tree/14.0/sale_order_weight

[IMP] Configuration to display the weight or not on the sale order report.

Configuration is available for the total order, and the order on each lines.

[REM] remove 'delivered' weight for the following reasons:

- The delivered quantity on sale order lines is a non stored field.
  So, the recompute of the field fails,
  and the field total_delivered_weight is not updated correctly.
- It is legit to want to have only Ordered weight
  and not Delivered Weight, and the addition of the delivered weight
  make the database bigger, and the process slower.

It is so Better to create a dedicated module for delivered weight