[FIX] sale_order_invoicing_finished_task: Change methods that do not exist and/or should not be used
Changes done:
- Change the _get_to_invoice_qty() method (no longer exists) to _compute_qty_to_invoice()
- Do not use the _get_invoiced() method; replace it with _compute_invoice_status ()
- Make tests more readable + use Form() to avoid manually calling onchange methods

Although the _get_invoiced() method exists, the name is confusing; it is actually a compute. The appropriate and most readable approach is to use the appropriate invoice_status value in each line of the order, and the order status will update itself.

Similar to https://github.com/OCA/sale-workflow/pull/3838

TT57321